Silence int/uint conversion warnings

+++ src/occsimplifier.cpp
@@ -949,7 +949,7 @@ bool OccSimplifier::eliminate_vars()
 
     while(varelim_num_limit > 0
         && *limit_to_decrease > 0
-        && grow < solver->conf.min_bva_gain
+        && grow < (uint32_t)solver->conf.min_bva_gain
         //&& grow < 1 //solver->conf.min_bva_gain
     ) {
         if (solver->conf.verbosity) {
